+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.15.3](https://github.com/RonenMars/threadbase-streamer/compare/v1.15.2...v1.15.3) (2026-06-19)
+
+### Bug Fixes
+
+* **server:** broadcast existing JSONL lines immediately on watcher wire ([af6ddce](https://github.com/RonenMars/threadbase-streamer/commit/af6ddcefb289947daf67dd13c99594e98ccadd07))
+
+## [1.15.2](https://github.com/RonenMars/threadbase-streamer/compare/v1.15.1...v1.15.2) (2026-06-19)
+
+### Bug Fixes
+
+* **server:** fallback to recently-modified JSONL when session UUID differs on resume ([8011ea2](https://github.com/RonenMars/threadbase-streamer/commit/8011ea2c41c4ba7412af5eb3b81dc60ee69b85d1))
+
 ## [1.15.1](https://github.com/RonenMars/threadbase-streamer/compare/v1.15.0...v1.15.1) (2026-06-18)
 
 ### Performance Improvements